引数の型変更

+++ src/main/java/client/Calculator.java
@@ -1,7 +1,7 @@
 package example;
 
 public class Calculator {
-    public int add(int a, int b) {
+    public int add(int a, double b) {
         return a + b;
     }
